backend/requirements.txt
fastapi
uvicorn
neo4j
python-dotenv
anthropic
sentence-transformers
pydantic